+++ auto_operation/visualizer/contexts/index.ts
@@ -1,0 +1,4 @@
+import { createContext } from "react";
+import { Blocks } from "../types";
+
+export const BlocksContext = createContext<Blocks>({});
